# docs/api-doc/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.2
  title: api.ramassage.epitech.eu
  description: |
    Documentation expliquant le fonctionnement de l'api de ramassage.epitech.eu
host: 'api.ramassage.epitech.eu'
basePath: /1.0
paths:
...
  '/user/{user_id}':
    get:
      tags:
        - user
      description: |
        Récupérer l'objet `User` correspondant à l'**userId**.
      produces:
        - application/json
      parameters:
        - in: path
          name: user_id
          description: "ID ou login de l'utilisateur"
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/User'
        '404':
          description: 'User #{userId} not found'
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - user
      description: |
        Modifier un  `User`.
      produces:
        - application/json
      parameters:
        - in: path
          name: user_id
          description: "ID de l'utilisateur uniquement"
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/User'
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
    patch:
      tags:
        - user
      description: |
        Modifier un champ de `User`.
      produces:
        - application/json
      parameters:
        - in: path
          name: user_id
          description: "ID de l'utilisateur uniquement"
          required: true
          type: integer
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/User'
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - user
      description: |
        Supprimer un `User`.
      produces:
        - application/json
      parameters:
        - in: path
          name: user_id
          description: "ID de l'utilisateur uniquement"
          required: true
          type: integer
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '404':
          description: User not found
          schema:
            $ref: '#/definitions/Error'
...
  '/template/{template_id}':
    get:
      tags:
        - template
      description: |
        Récupérer l'objet `Template` correspondant à l'**template_id**.
      produces:
        - application/json
      parameters:
        - in: path
          name: template_id
          description: ID du Template
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Template'
        '404':
          description: 'Template #{template_id} not found'
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - template
      description: |
        Modifier un  `Template`.
      produces:
        - application/json
      parameters:
        - in: path
          name: template_id
          description: ID du Template uniquement
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Template'
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Template'
        '404':
          description: Template not found
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
    patch:
      tags:
        - template
      description: |
        Modifier un champ de `Template`.
      produces:
        - application/json
      parameters:
        - in: path
          name: template_id
          description: ID du Template uniquement
          required: true
          type: integer
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Template'
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Template'
        '404':
          description: Template not found
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
...
  '/task/project/{project_id}':
    get:
      tags:
        - task
      description: |
        Récupérer les `Task` liées au Project correspondant à **project_id**.
      produces:
        - application/json
      parameters:
        - in: path
          name: project_id
          description: ID du projet
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Task'
        '404':
          description: 'Project #{project_id} not found'
          schema:
            $ref: '#/definitions/Error'
  /project/:
    get:
      tags:
        - project
      description: >
        Récupérer les `Project` en cours (où la date de fin se situe au delà de
        maintenant).
      produces:
        - application/json
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
    post:
      tags:
        - project
      description: |
        Créer un nouveau `Project`.
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Project'
      security:
        - api_signature: []
      responses:
        '201':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
  '/project/{project_id}':
    get:
      tags:
        - project
      description: |
        Récupérer l'objet `Project` correspondant au **project_id**.
      produces:
        - application/json
      parameters:
        - in: path
          name: project_id
          description: ID du Project
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Project'
        '404':
          description: 'Project #{project_id} not found'
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - project
      description: |
        Modifier un  `Project`.
      produces:
        - application/json
      parameters:
        - in: path
          name: project_id
          description: ID du Project uniquement
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Project'
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        '404':
          description: Project not found
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
    patch:
      tags:
        - project
      description: |
        Modifier un champ de la `Project`.
      produces:
        - application/json
      parameters:
        - in: path
          name: project_id
          description: ID du Project uniquement
          required: true
          type: integer
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/Project'
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        '404':
          description: Project not found
          schema:
            $ref: '#/definitions/Error'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/Error'
  '/project/slug/{slug}':
    get:
      tags:
        - project
      description: |
        Récupérer les `Project` liés au **slug**.
      produces:
        - application/json
      parameters:
        - in: path
          name: slug
          description: Slug correspondant au Template
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        '404':
          description: 'Template #{slug} not found'
          schema:
            $ref: '#/definitions/Error'
  '/project/template/{template_id}':
    get:
      tags:
        - project
      description: |
        Récupérer les `Project` liés au **template_id**.
      produces:
        - application/json
      parameters:
        - in: path
          name: template_id
          description: ID correspondant au Template
          required: true
          type: integer
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        '404':
          description: 'Template #{slug} not found'
          schema:
            $ref: '#/definitions/Error'
  '/project/token/{token}':
    get:
      tags:
        - project
      description: |
        Récupérer l'objet `Project` correspondant au **token**.
      produces:
        - application/json
      parameters:
        - in: path
          name: token
          description: Token Project
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Project'
        '404':
          description: 'Project #{token} not found'
          schema:
            $ref: '#/definitions/Error'
  /project/past:
    get:
      tags:
        - project
      description: >
        Récupérer les `Project` passés (où la date de fin se situe avant
        aujourd'hui).
      produces:
        - application/json
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
  /project/all:
    get:
      tags:
        - project
      description: |
        Récupérer tous les `Project`.
      produces:
        - application/json
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
  '/project/user/{login}':
    get:
      tags:
        - project
      description: |
        Récupérer les `Project` courant de l'étudiant **login**.
      produces:
        - application/json
      parameters:
        - in: path
          name: login
          description: "Login de l'étudiant"
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        '404':
          description: 'User #{login} not found'
          schema:
            $ref: '#/definitions/Error'
  '/project/all/user/{login}':
    get:
      tags:
        - project
      description: |
        Récupérer tous les `Project` de l'étudiant **login**.
      produces:
        - application/json
      parameters:
        - in: path
          name: login
          description: "Login de l'étudiant"
          required: true
          type: string
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
        '404':
          description: 'User #{login} not found'
          schema:
            $ref: '#/definitions/Error'
  '/project/{project_id}/notes':
    post:
      tags:
        - project
      description: >
        Upload des notes pour l'objet `Project` correspondant au **project_id**.
      produces:
        - application/json
      parameters:
        - in: path
          name: project_id
          description: ID du Project
          required: true
          type: string
        - in: body
          name: body
          description: "notes en tant qu'object json ou bien en tant que csv avec header (mini: login:note;comment)"
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Notes'
      security:
        - api_signature: []
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/Project'
        '404':
          description: 'Project #{project_id} not found'
          schema:
            $ref: '#/definitions/Error'
...
securityDefinitions:
  api_signature:
    type: apiKey
    name: Authorization
    in: header
definitions:
  Error:
    properties:
      status:
        type: integer
      title:
        type: string
      detail:
        type: string
        description: "Message plus explicite de l'erreur"
      id:
        type: string
      code:
        type: integer
  User:
    required:
      - firstname
      - lastname
      - login
    properties:
      id:
        type: integer
      firstname:
        type: string
      lastname:
        type: string
      login:
        type: string
  Notes:
    properties:
      login:
        type: string
      note:
        type: string
      comment:
        type: string
  Template:
    properties:
      id:
        type: integer
      codemodule:
        type: string
      slug:
        type: string
      repository_name:
        type: string
        description: Nom du dépôt à ramasser
      call_moulitriche:
        type: boolean
        description: Appel au webservice de triche ?
      call_judge:
        type: boolean
        description: Correction?
      judge_uri:
        type: string
        description: login@machine_de_correction
      judge_rule:
        type: string
        description: "'Rule' identifiant unique coté correction"
      judge_preliminary_exec:
        type: string
        description: Commande de correction pour les préliminaires
      judge_final_exec:
        type: string
        description: Commande de correction pour le rendu final
      school:
        type: string
        description: ''
...
  Task:
    properties:
      id:
        type: integer
      type:
        type: string
        description: 'Enum("auto", "preliminary", "manual")'
      launch_date:
        type: string
        format: date-time
        description: Datetime mysql format
      status:
        type: string
        description: 'Enum("ongoing", "todo", "succeed", "failed")'
      project_id:
        type: integer
      project:
        $ref: '#/definitions/Project'
      exec_cmd:
        type: string
      extend:
        type: string
        description: Json dump
...